fix(dashboard): state-neutral pre-mount cooldown placeholder for tripped breakers
Cursor flagged that a TRIPPED breaker whose on-chain cooldown had already
elapsed showed a false "cooldown active"/✗ pre-mount, because
cooldownRemainingSecFrom returns null (unresolved) rather than 0 (elapsed)
before the client clock mounts. Fix: the pre-mount placeholder is now
state-neutral ("—", never "active" or "elapsed") in both ThresholdMetric
and ResetPathBanner, resolving to the real active/elapsed state once the
ticker's first tick lands. Extended the existing hydrateRoot determinism
test with an elapsed-cooldown fixture pinning the fix.

Also, while in this territory:
- claude[bot] flagged breaker-panel.tsx over the repo's 600-line soft cap;
  extracted the pure Fixidity math / cooldown / trip-count helpers and the
  BreakerPresentation view model into a sibling breaker-panel-math.ts
  (verbatim move), bringing the component file from 732 to 489 lines.
- Codex flagged the last-trip anchor's title still using formatTimestamp
  (toLocaleString) even though BreakerPanel now SSR-renders with
  fallbackData; switched to the existing useSsrSafeTimestamp hook so the
  title is deterministic pre-mount too, and extended the hydration test to
  pin the UTC-then-local transition.
- Codex flagged that MarketHoursPill's weekend reopen countdown disappears
  for a pre-mounted, SSR-prefetched breaker closure and then reappears
  post-mount, reintroducing the #1222 pill-width/wrap shift. Reserved the
  countdown slot with the same state-neutral "—" pattern
  (deriveMarketHoursState now exposes showCountdown/countdownText), and
  added a hydrateRoot test pinning the neutral-placeholder → real-countdown
  transition.
- Dropped pool-header.tsx's stale "tracked as a follow-up in issue #1238"
  comment now that #1238 shipped in #1258.
