+++ ui-dashboard/src/components/breaker-panel.tsx
@@ -3,18 +3,32 @@
 import { useEffect, useState } from "react";
 import { useGQL } from "@/lib/graphql";
 import { useNetwork } from "@/components/network-provider";
-import {
-  POOL_BREAKER_CONFIG,
-  type PoolBreakerConfigResponse,
-} from "@/lib/queries";
+import { type PoolBreakerConfigResponse } from "@/lib/queries";
 import type { BreakerConfig, BreakerTripEvent, Pool } from "@/lib/types";
 import { isVirtualPool } from "@/lib/types";
 import { effectiveBreakerThreshold, pickTrippableConfig } from "@/lib/breaker";
 import { Tooltip } from "@/components/tooltip";
 import { explorerTxUrl } from "@/lib/tokens";
-import { formatTimestamp } from "@/lib/format";
 import { formatDurationShort } from "@/lib/bridge-status";
-import { useNowSeconds, useSsrSafeRelative } from "@/hooks/use-now-seconds";
+import {
+  useNowSeconds,
+  useSsrSafeRelative,
+  useSsrSafeTimestamp,
+} from "@/hooks/use-now-seconds";
+// Fixidity math, cooldown/trip-count derivations, and the presentation view
+// model live in the sibling module to keep this file under the repo's
+// file-size soft cap — see breaker-panel-math.ts's header comment.
+import {
+  breakerConfigQuery,
+  breakerPresentation,
+  computeLiveDelta,
+  cooldownRemainingSecFrom,
+  countTripsToday,
+  deltaBarStyle,
+  effectiveCooldown,
+  isBreakerConfigQueryPending,
+  type BreakerPresentation,
+} from "./breaker-panel-math";
 
 type Props = {
   pool: Pool;
@@ -28,243 +42,6 @@ type Response = {
   BreakerConfig: BreakerConfig[];
   BreakerTripEvent: BreakerTripEvent[];
 };
-
-const FIXED_1 = BigInt(10) ** BigInt(24);
-// Breaker thresholds are stored as Fixidity (1e24 = 100%). Keep one decimal
-// pair throughout the panel: 4.00% / 0.150% etc. We render two decimals for
-// FX (4.00%) and three for stablecoin (0.150%) to match the precision of
-// the on-chain config without trailing-zero clutter.
-const PRECISION_BY_KIND: Record<string, number> = {
-  MEDIAN_DELTA: 2,
-  VALUE_DELTA: 3,
-};
-
-function formatFixidityPct(
-  raw: string | null | undefined,
-  precision: number,
-): string | null {
-  if (raw === null || raw === undefined) return null;
-  // Convert Fixidity (1e24=100%) to percent. Avoid Number for big values —
-  // do it in BigInt with a /1e22 trick to preserve precision. Treats 0 as a
-  // legitimate value (median exactly equals reference) — renders "0.00%",
-  // not the missing-data dash.
-  const value = BigInt(raw);
-  if (value < BigInt(0)) return null;
-  const scale = BigInt(10) ** BigInt(22 - precision);
-  const scaled = value / scale;
-  const whole = scaled / BigInt(10) ** BigInt(precision);
-  const frac = scaled % BigInt(10) ** BigInt(precision);
-  return `${whole}.${frac.toString().padStart(precision, "0")}%`;
-}
-
-/** Format a Fixidity-scaled number with 6 decimals (e.g. EMA = 1.171560). */
-function formatFixidityValue(raw: string | null | undefined): string | null {
-  if (!raw) return null;
-  const value = BigInt(raw);
-  if (value === BigInt(0)) return null;
-  const whole = value / FIXED_1;
-  const frac = value % FIXED_1;
-  // Render 6 decimals: scale frac by 10^6 / FIXED_1 = 10^6 / 10^24 = 10^-18.
-  const fracScaled = frac / BigInt(10) ** BigInt(18);
-  return `${whole}.${fracScaled.toString().padStart(6, "0")}`;
-}
-
-function fixidityOrNull(raw: string | null | undefined): bigint | null {
-  if (raw == null) return null;
-  const value = BigInt(raw);
-  return value === BigInt(0) ? null : value;
-}
-
-function breakerConfigQuery(
-  isVirtual: boolean,
-  rateFeedID: string,
-): string | null {
-  return !isVirtual && rateFeedID ? POOL_BREAKER_CONFIG : null;
-}
-
-// When the SSR prefetch supplies `fallbackData` (issue #1237), `data` is
-// populated on first paint and this is false — the panel renders its resolved
-// shape immediately (full strip or null, no shimmer). This stays true only on
-// the degraded path (prefetch missed / partial) while the client query is
-// genuinely in flight, so BreakerPanel renders a matching-shape shimmer instead
-// of the null→content jump (issue #1222's measured +119px header jump).
-function isBreakerConfigQueryPending(
-  isVirtual: boolean,
-  rateFeedID: string,
-  data: unknown,
-  isLoading: boolean,
-): boolean {
-  return !isVirtual && !!rateFeedID && data === undefined && isLoading;
-}
-
-/** Effective cooldown in seconds. Per-feed override else breaker default. */
-function effectiveCooldown(cfg: BreakerConfig): bigint {
-  const override = BigInt(cfg.cooldownTime);
-  if (override > BigInt(0)) return override;
-  return BigInt(cfg.breaker.defaultCooldownTime);
-}
-
-/** Seconds until `cooldownEndsAt`, or `null` while the SSR-safe ticker
- *  hasn't read a real wall clock yet (`now === null` pre-mount — see
- *  BreakerPanel's `now` state). Reading `Date.now()` directly at render time
- *  would diverge between the server pass and the client's hydration pass now
- *  that fallbackData paints a TRIPPED breaker's real content on first paint
- *  (issue #1237 round 2); `null` renders a deterministic "cooldown active"
- *  placeholder in ThresholdMetric/ResetPathBanner until the ticker's first
- *  tick lands. */
-function cooldownRemainingSecFrom(
-  cooldownEndsAt: number,
-  now: number | null,
-): number | null {
-  return now === null ? null : Math.max(0, cooldownEndsAt - now);
-}
-
-/** |median - reference| / reference, returned as a Fixidity ratio (1e24 = 100%).
- * Returns null if either input is missing OR is the on-chain `0` sentinel:
- * SortedOracles returns rate `0` when all oracle reports have expired,
- * `medianRatesEMA = 0` is the contract's "uninitialized" marker (until the
- * first MedianUpdated seeds it), and a `referenceValue` of `0` would be a
- * mis-set peg (also produces a divide-by-zero). In all three cases the live
- * Δ is meaningless, so render the missing-data dash. */
-function computeLiveDelta(cfg: BreakerConfig): bigint | null {
-  const median = fixidityOrNull(cfg.lastMedianRate);
-  const reference =
-    cfg.breaker.kind === "MEDIAN_DELTA"
-      ? fixidityOrNull(cfg.medianRatesEMA)
-      : fixidityOrNull(cfg.referenceValue);
-  if (median == null || reference == null) return null;
-  const diff = median > reference ? median - reference : reference - median;
-  return (diff * FIXED_1) / reference;
-}
-
-function referenceValue(cfg: BreakerConfig): bigint | null {
-  const raw =
-    cfg.breaker.kind === "MEDIAN_DELTA"
-      ? cfg.medianRatesEMA
-      : cfg.referenceValue;
-  return fixidityOrNull(raw);
-}
-
-function actualValue(cfg: BreakerConfig): bigint | null {
-  return fixidityOrNull(cfg.lastMedianRate);
-}
-
-function valueDeltaDirection(cfg: BreakerConfig): "above" | "below" | null {
-  const reference = referenceValue(cfg);
-  const actual = actualValue(cfg);
-  if (reference == null || actual == null || actual === reference) return null;
-  return actual > reference ? "above" : "below";
-}
-
-type BreakerPresentation = {
-  referenceLabel: string;
-  liveDeltaLabel: string;
-  thresholdCaption: string;
-  formattedThreshold: string;
-  formattedCooldown: string;
-  formattedReference: string | null;
-  formattedActual: string | null;
-  formattedLiveDelta: string;
-  breachedValueClass: string;
-  referenceCaption: string;
-  isOverTolerance: boolean;
-};
-
-function isMedianDelta(cfg: BreakerConfig): boolean {
-  return cfg.breaker.kind === "MEDIAN_DELTA";
-}
-
-function referenceCaptionFor(
-  cfg: BreakerConfig,
-  tripped: boolean,
-  isOverTolerance: boolean,
-  formattedLiveDelta: string,
-): string {
-  if (isMedianDelta(cfg)) {
-    return `smoothing ${formatFixidityPct(cfg.smoothingFactor, 1) ?? "—"}`;
-  }
-  const pegDirection = valueDeltaDirection(cfg);
-  if (tripped && isOverTolerance && pegDirection) {
-    // `isOverTolerance` implies a non-null live delta, so this is never "—".
-    return `${formattedLiveDelta} ${pegDirection} peg`;
-  }
-  return "fixed peg";
-}
-
-function breakerPresentation(
-  cfg: BreakerConfig,
-  threshold: bigint,
-  cooldown: bigint,
-  liveDelta: bigint | null,
-  tripped: boolean,
-): BreakerPresentation {
-  const kind = cfg.breaker.kind;
-  const precision = PRECISION_BY_KIND[kind] ?? 2;
-  const formattedThreshold =
-    formatFixidityPct(threshold.toString(), precision) ?? "—";
-  const formattedLiveDelta =
-    liveDelta != null
-      ? (formatFixidityPct(liveDelta.toString(), precision) ?? "—")
-      : "—";
-  const isOverTolerance =
-    threshold > BigInt(0) && liveDelta != null && liveDelta >= threshold;
-
-  return {
-    referenceLabel:
-      kind === "MEDIAN_DELTA"
-        ? "EMA Reference vs Actual"
-        : "Reference vs Actual",
-    liveDeltaLabel:
-      kind === "MEDIAN_DELTA"
-        ? "Δ Oracle Price vs EMA"
-        : "Δ Oracle Price vs Peg",
-    thresholdCaption:
-      kind === "MEDIAN_DELTA"
-        ? `trips at >${formattedThreshold} from EMA`
-        : `trips at >${formattedThreshold} from peg`,
-    formattedThreshold,
-    formattedCooldown: formatDurationShort(Number(cooldown)),
-    formattedReference: formatFixidityValue(
-      kind === "MEDIAN_DELTA" ? cfg.medianRatesEMA : cfg.referenceValue,
-    ),
-    formattedActual: formatFixidityValue(cfg.lastMedianRate),
-    formattedLiveDelta,
-    breachedValueClass:
-      tripped && isOverTolerance ? "text-red-300" : "text-white",
-    referenceCaption: referenceCaptionFor(
-      cfg,
-      tripped,
-      isOverTolerance,
-      formattedLiveDelta,
-    ),
-    isOverTolerance,
-  };
-}
-
-/** Returns the bar fill (0-100) and color class for the live-Δ bar. Mirrors
- * the deviation-bar conventions in components/pool-header/deviation-cell.tsx. */
-function deltaBarStyle(
-  deltaFixidity: bigint,
-  thresholdFixidity: bigint,
-): {
-  pct: number;
-  color: string;
-} {
-  if (thresholdFixidity <= BigInt(0)) {
-    return { pct: 0, color: "bg-slate-600" };
-  }
-  // ratio = delta / threshold, capped at 1.5 for visual purposes.
-  const ratioBP = (deltaFixidity * BigInt(10000)) / thresholdFixidity;
-  const ratio = Number(ratioBP) / 10000;
-  const pct = Math.min(ratio * 100, 100);
-  const color =
-    ratio >= 1
-      ? "bg-red-500"
-      : ratio >= 0.8
-        ? "bg-yellow-500"
-        : "bg-emerald-500";
-  return { pct, color };
-}
 
 function BreakerIdentityMetric({
   cfg,
@@ -339,10 +116,11 @@ function ThresholdMetric({
   const cooldownActive =
     tripped && cooldownRemainingSec !== null && cooldownRemainingSec > 0;
   // `cooldownRemainingSec` is `null` pre-mount even when `tripped` (the
-  // ticker hasn't read the wall clock yet — see BreakerPanel's `now` state)
-  // — render a deterministic caption instead of an exact duration so the
-  // server and hydration renders agree; the countdown itself lands after
-  // mount.
+  // ticker hasn't read the wall clock yet — see BreakerPanel's `now` state).
+  // We can't tell from `null` whether the on-chain cooldown is still active
+  // or has already elapsed, so render the state-neutral "—" rather than
+  // asserting either one; the real caption lands once the ticker's first
+  // tick resolves `now`.
   const cooldownPending = tripped && cooldownRemainingSec === null;
   return (
     <div>
@@ -357,7 +135,7 @@ function ThresholdMetric({
           {cooldownActive && cooldownRemainingSec !== null
             ? `${formatDurationShort(cooldownRemainingSec)} left`
             : cooldownPending
-              ? "cooldown active"
+              ? "—"
               : presentation.thresholdCaption}
         </span>
       </dd>
@@ -418,15 +196,17 @@ function LastTripMetric({
   tripsToday: number;
 }): React.ReactElement {
   const lastTripTs = cfg.lastTripAt;
-  // SSR-safe relative label (mirrors the header's `createdRelative` pattern,
+  // SSR-safe relative label + title (mirrors the header's `createdRelative`
+  // pattern and rebalance-status-value.tsx's `LastRebalanceSubtitle`,
   // hooks/use-now-seconds.ts): the SSR prefetch's fallbackData now paints
   // this panel's real content on first paint for pools with a trip-able
-  // breaker (issue #1237), so a plain `relativeTime(lastTripTs)` read at
-  // render time could disagree between the page's ISR-cached bake time and
-  // the viewer's hydration clock. `useSsrSafeRelative` renders a
-  // deterministic UTC date on the server + hydration render, then the live
-  // "N ago" label after mount.
+  // breaker (issue #1237), so a plain `relativeTime`/`formatTimestamp` read
+  // at render time could disagree between the page's ISR-cached bake time
+  // (UTC) and the viewer's hydration clock (local tz/locale). Both hooks
+  // render a deterministic UTC value on the server + hydration render, then
+  // the live locale-formatted value after mount.
   const lastTripRelative = useSsrSafeRelative(lastTripTs);
+  const lastTripTitle = useSsrSafeTimestamp(lastTripTs);
   return (
     <div>
       <dt className="text-slate-400 flex items-center justify-between gap-1">
@@ -446,7 +226,7 @@ function LastTripMetric({
             className={`${
               tripped ? "text-red-300" : "text-indigo-300"
             } hover:text-indigo-400`}
-            title={formatTimestamp(lastTripTs)}
+            title={lastTripTitle}
           >
             {lastTripRelative}
           </a>
@@ -465,38 +245,52 @@ function LastTripMetric({
 }
 
 function ResetPathBanner({
-  cooldownElapsed,
   cooldownRemainingSec,
   rateInBand,
   liveDelta,
   presentation,
 }: {
-  cooldownElapsed: boolean;
   cooldownRemainingSec: number | null;
   rateInBand: boolean;
   liveDelta: bigint | null;
   presentation: BreakerPresentation;
 }): React.ReactElement {
+  // `cooldownRemainingSec === null` pre-mount (see BreakerPanel's `now`
+  // state) doesn't tell us whether the on-chain cooldown is still active or
+  // has already elapsed — asserting either would misstate reset readiness
+  // for a TRIPPED breaker whose cooldown already ended before mount. Render
+  // a state-neutral "—" instead of guessing ✓/✗; the real state lands once
+  // the ticker's first tick resolves `now`.
+  const cooldownElapsed = cooldownRemainingSec === 0;
   return (
     <div className="mt-4 rounded border border-red-900/40 bg-red-950/30 px-3 py-2 flex flex-wrap items-center gap-x-4 gap-y-1.5 text-xs">
       <span className="text-red-300 font-medium">Reset path</span>
       <span className="inline-flex items-center gap-1">
-        <span className={cooldownElapsed ? "text-emerald-300" : "text-red-300"}>
-          {cooldownElapsed ? "✓" : "✗"}
+        <span
+          className={
+            cooldownRemainingSec === null
+              ? "text-slate-500"
+              : cooldownElapsed
+                ? "text-emerald-300"
+                : "text-red-300"
+          }
+        >
+          {cooldownRemainingSec === null ? "—" : cooldownElapsed ? "✓" : "✗"}
         </span>
         Cooldown
         <span
           className={`font-mono ${
-            cooldownElapsed ? "text-emerald-300" : "text-amber-300"
+            cooldownRemainingSec === null
+              ? "text-slate-500"
+              : cooldownElapsed
+                ? "text-emerald-300"
+                : "text-amber-300"
           }`}
         >
-          {cooldownElapsed
-            ? "elapsed"
-            : cooldownRemainingSec === null
-              ? // Pre-mount (see BreakerPanel's `now` state): deterministic
-                // placeholder instead of a duration that could disagree
-                // between the server and hydration renders.
-                "cooldown active"
+          {cooldownRemainingSec === null
+            ? "—"
+            : cooldownElapsed
+              ? "elapsed"
               : formatDurationShort(cooldownRemainingSec)}
         </span>
       </span>
@@ -524,28 +318,6 @@ function ResetPathBanner({
       </span>
     </div>
   );
-}
-
-/** Number of trips for `breakerAddress` since UTC midnight. `todayNowSeconds
- *  === null` (server + hydration render, see useNowSeconds) deterministically
- *  returns 0 rather than risking a UTC-midnight mismatch between the page's
- *  bake time and the viewer's clock — the real count settles in after mount
- *  (issue #1237). Filters by breaker address — the query is feed-scoped, but
- *  a single feed could surface multiple breakers' trips (only one trip-able
- *  breaker today, but MarketHours-style additions later would drift the
- *  count from cfg.tripCountLifetime if we didn't scope it). */
-function countTripsToday(
-  todayNowSeconds: number | null,
-  trips: BreakerTripEvent[],
-  breakerAddress: string,
-): number {
-  if (todayNowSeconds === null) return 0;
-  const todayMidnightSec = Math.floor(todayNowSeconds / 86400) * 86400;
-  return trips.filter(
-    (t) =>
-      Number(t.blockTimestamp) >= todayMidnightSec &&
-      t.breaker.address === breakerAddress,
-  ).length;
 }
 
 export function BreakerPanel({
@@ -639,8 +411,9 @@ export function BreakerPanel({
     cfg.breaker.address,
   );
 
-  // Reset-path conditions (mirror BreakerBox.tryResetBreaker).
-  const cooldownElapsed = cooldownRemainingSec === 0;
+  // Reset-path condition (mirror BreakerBox.tryResetBreaker). Cooldown's own
+  // elapsed/pending state is derived inside ResetPathBanner from
+  // cooldownRemainingSec directly (see its comment).
   const rateInBand = liveDelta != null && liveDelta < threshold;
 
   return (
@@ -670,7 +443,6 @@ export function BreakerPanel({
 
       {tripped && (
         <ResetPathBanner
-          cooldownElapsed={cooldownElapsed}
           cooldownRemainingSec={cooldownRemainingSec}
           rateInBand={rateInBand}
           liveDelta={liveDelta}
